requirements.txt
# Core FastAPI
fastapi==0.121.0
uvicorn[standard]==0.38.0
python-multipart==0.0.20

# Database
SQLAlchemy==2.0.41
alembic==1.17.1
argon2-cffi=25.1.0
argon2-cffi-bindings=

# Authentication
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==5.0.0
argon2-cffi = 25.1.0

# LangChain & AI
langchain==1.0.3
langgraph==1.0.3
langchain-groq==1.0.1
langchain-community==0.4.1

# Validation
pydantic==2.11.7
email-validator==2.3.0

# Environment
python-dotenv==1.1.1

# Tools
duckduckgo-search==8.1.1
requests==2.32.5

# Testing
pytest==9.0.1
httpx==0.28.1
